Add files via upload
Here we create data and using that data to show use who are just getting started in the data analysis or scientist field how this library can be use for different function. Here is just one of it function. This is Plotly used for creating Graph, and other application such as a dashboard but i created a heatwave here so see and run my code to get an idea of you how can use this library. Happy coding, dude :-)